Moved registration form to LoginPage.jsx and created HomePage.jsx; created UserContext.jsx
