+++ src/pages/HomePage.jsx
@@ -1,0 +1,35 @@
+import {useContext} from "react";
+import {UserContext} from "../contexts/UserContext";
+
+function HomePage() {
+  const {userInfo, setUserInfo} = useContext(UserContext);
+
+  return (
+    <>
+      <p>
+        New learner user registration was succesful.
+      </p>
+
+      <table border={2}>
+        <tbody>
+          <tr>
+            <th>UID</th><td>{userInfo.userUID}</td>
+          </tr>
+          <tr>
+            <th>Name</th><td>{userInfo.name}</td>
+          </tr>
+          <tr>
+            <th>E-mail</th><td>{userInfo.email}</td>
+          </tr>
+        </tbody>
+      </table>
+      <br />
+
+      <button onClick={() => setUserInfo(null)}>
+        Go back to login screen.
+      </button>
+    </>
+  )
+}
+
+export default HomePage;
